filteredgroupapi: Add Region field and WithRegion ctx
